a bit of codeclimate
