+++ app/src/main/java/ch/epfl/polychef/fragments/OnlineMiniaturesFragment.java
@@ -35,6 +35,7 @@ import ch.epfl.polychef.users.UserStorage;
 import ch.epfl.polychef.utils.MiniatureAdapter;
 import ch.epfl.polychef.utils.Preconditions;
 import ch.epfl.polychef.utils.RecipeMiniatureAdapter;
+import ch.epfl.polychef.utils.Similarity;
 
 public class OnlineMiniaturesFragment extends Fragment implements CallHandler<List<Miniatures>>{
 
@@ -64,6 +65,28 @@ public class OnlineMiniaturesFragment extends Fragment implements CallHandler<Li
     private RecipeStorage recipeStorage;
     private ImageStorage imageStorage;
     private UserStorage userStorage;
+
+    private Comparator<Miniatures> similarityComparator = (o1, o2) -> {
+        String s1;
+        String s2;
+        if(o1.getClass().equals(Recipe.class)){
+            s1 = ((Recipe)o1).getName();
+        }else{
+            s1 = ((User)o1).getUsername();
+        }
+        if(o2.getClass().equals(Recipe.class)){
+            s2 = ((Recipe)o2).getName();
+        }else{
+            s2 = ((User)o2).getUsername();
+        }
+        if(Similarity.similarity(s1,actualQuery)>Similarity.similarity(s2,actualQuery)){
+            return -1;
+        }else if(Similarity.similarity(s1,actualQuery)==Similarity.similarity(s2,actualQuery)){
+            return 0;
+        }else {
+            return 1;
+        }
+    };
 
     public OnlineMiniaturesFragment(){
         
@@ -182,29 +205,8 @@ public class OnlineMiniaturesFragment extends Fragment implements CallHandler<Li
     @Override
     public void onSuccess(List<Miniatures> data){
         if(isSearching){
-            Comparator<Miniatures> myComparator = (o1, o2) -> {
-                String s1;
-                String s2;
-                if(o1.getClass().equals(Recipe.class)){
-                    s1 = ((Recipe)o1).getName();
-                }else{
-                    s1 = ((User)o1).getUsername();
-                }
-                if(o2.getClass().equals(Recipe.class)){
-                    s2 = ((Recipe)o2).getName();
-                }else{
-                    s2 = ((User)o2).getUsername();
-                }
-                if(similarity(s1,actualQuery)>similarity(s2,actualQuery)){
-                    return -1;
-                }else if(similarity(s1,actualQuery)==similarity(s2,actualQuery)){
-                    return 0;
-                }else {
-                    return 1;
-                }
-            };
             searchList.addAll(data);
-            Collections.sort(searchList,myComparator);
+            Collections.sort(searchList,similarityComparator);
             onlineRecyclerView.getAdapter().notifyDataSetChanged();
             return;
         }
@@ -242,44 +244,5 @@ public class OnlineMiniaturesFragment extends Fragment implements CallHandler<Li
 
     public RecyclerView getRecyclerView(){
         return onlineRecyclerView;
-    }
-
-    public static double similarity(String s1, String s2) {
-        String longer = s1, shorter = s2;
-        if (s1.length() < s2.length()) {
-            longer = s2; shorter = s1;
-        }
-        int longerLength = longer.length();
-        if (longerLength == 0) { return 1.0;}
-        return (longerLength - editDistance(longer, shorter)) / (double) longerLength;
-
-    }
-
-
-    public static int editDistance(String s1, String s2) {
-        s1 = s1.toLowerCase();
-        s2 = s2.toLowerCase();
-
-        int[] costs = new int[s2.length() + 1];
-        for (int i = 0; i <= s1.length(); i++) {
-            int lastValue = i;
-            for (int j = 0; j <= s2.length(); j++) {
-                if (i == 0)
-                    costs[j] = j;
-                else {
-                    if (j > 0) {
-                        int newValue = costs[j - 1];
-                        if (s1.charAt(i - 1) != s2.charAt(j - 1))
-                            newValue = Math.min(Math.min(newValue, lastValue),
-                                    costs[j]) + 1;
-                        costs[j - 1] = lastValue;
-                        lastValue = newValue;
-                    }
-                }
-            }
-            if (i > 0)
-                costs[s2.length()] = lastValue;
-        }
-        return costs[s2.length()];
     }
 }
